part 4 done

+++ part4/blog/controllers/blogs.js
@@ -1,14 +1,23 @@
 const blogsRouter = require('express').Router()
-const blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = (await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+    )
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(400).json({ error: 'user missing or not valid' })
+  }
 
   if (!body.title || !body.url) {
     return response.status(400).json({
@@ -20,16 +29,36 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  
+  const user = request.user
+
+  if (!user) {
+    return response.status(400).json({ error: 'user missing or not valid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+
+    user.blogs = user.blogs.filter(id => id.toString() !== request.params.id)
+
+    await user.save()
+
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
